backend/requirements.txt
fastapi
uvicorn
numpy
pillow
opencv-python
torch
torchvision
detectron2
pydantic
